move run, config to relevant modules
