+++ app/scripts/modules/core/cache/caches.module.js
@@ -2,10 +2,14 @@
 
 let angular = require('angular');
 
-module.exports = angular.module('spinnaker.core.cache', [
+module.exports = angular
+  .module('spinnaker.core.cache', [
     require('./deckCacheFactory.js'),
     require('./cacheInitializer.js'),
     require('./collapsibleSectionStateCache.js'),
     require('./infrastructureCaches.js'),
   ])
+  .run(function(cacheInitializer) {
+    cacheInitializer.initialize();
+  })
   .name;
