+++ app/scripts/modules/core/modal/modal.module.js
@@ -4,7 +4,8 @@ let angular = require('angular');
 
 require('./modals.less');
 
-module.exports = angular.module('spinnaker.core.modal', [
+module.exports = angular
+  .module('spinnaker.core.modal', [
     require('./modalOverlay.directive.js'),
     require('./modalPage.directive.js'),
     require('./buttons/modalClose.directive.js'),
@@ -12,5 +13,10 @@ module.exports = angular.module('spinnaker.core.modal', [
     require('./wizard/modalWizard.directive.js'),
     require('./wizard/modalWizard.service.js'),
     require('./wizard/wizardPage.directive.js'),
-  ]
-).name;
+  ]).run(function($rootScope, $uibModalStack) {
+    $rootScope.$on('$stateChangeStart', function(event, toState, toParams, fromState, fromParams) {
+      if (!fromParams.allowModalToStayOpen) {
+        $uibModalStack.dismissAll();
+      }
+    });
+  }).name;
