+++ app/scripts/modules/core/utils/utils.module.js
@@ -10,5 +10,6 @@ module.exports = angular.module('spinnaker.utils', [
   require('./rx.js'),
   require('./uuid.service.js'),
   require('./appendTransform.js'),
-  require('./clipboard/copyToClipboard.directive.js')
+  require('./clipboard/copyToClipboard.directive.js'),
+  require('./timeFormatters.js'),
 ]).name;
